python/pub-crawl-map/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
GDAL==3.8.4
psycopg2==2.9.11
sqlparse==0.5.4
